Refactor so text64 and test80color link cleanly.
Basically want to be able to use text80 without bringing in text80color.
... and likewise for text64.

Should be no need to rebuild to optionally use text64 or f18a based 80 column color.
