+++ conio_getvram.c
@@ -8,10 +8,8 @@ unsigned int conio_getvram() {
         return VDP_SCREEN_TEXT(conio_y,conio_x)+gImage;
     } else if (nTextEnd==1919) {
         return VDP_SCREEN_TEXT80(conio_y,conio_x)+gImage;
-#ifdef ENABLE_TEXT64
     } else if (nTextRow == 64*23) {
         return VDP_SCREEN_TEXT64(conio_y,conio_x)+gImage;
-#endif
     } else {
         return VDP_SCREEN_POS(conio_y,conio_x)+gImage;
     }
