+++ vdp_bigbuf.c
@@ -1,0 +1,5 @@
+#include "vdp.h"
+
+// used by fast_scrn_scroll_default and fast_scrn_scroll64
+unsigned char vdp_bigbuf[256];
+
