+++ vdp_scrnscroll.c
@@ -9,16 +9,16 @@ void scrn_scroll() {
 		vdpmemcpy(adr-nLine, x, 4);
 	}
 	vdpmemset(nTextRow, ' ', nLine);	// clear the last line
-#ifdef ENABLE_F18A
-    if (gColor == 0x800) {
-	// copy color attributes the same way
-	extern unsigned int conio_scrnCol; // conio_bgcolor.c
 
-	for (int adr=nLine+gColor; adr<nTextEnd+gColor; adr+=4) {
-		vdpmemread(adr, x, 4);
-		vdpmemcpy(adr-nLine, x, 4);
+        // color attribute table is only at this address in 80 column color mode.
+	if (gColor == 0x800) {
+		// copy color attributes the same way
+		extern unsigned int conio_scrnCol; // conio_bgcolor.c
+
+		for (int adr=nLine+gColor; adr<nTextEnd+gColor; adr+=4) {
+			vdpmemread(adr, x, 4);
+			vdpmemcpy(adr-nLine, x, 4);
+		}
+		vdpmemset(nTextRow+gColor, conio_scrnCol, nLine);	// clear the last line
 	}
-	vdpmemset(nTextRow+gColor, conio_scrnCol, nLine);	// clear the last line
-    }
-#endif
 }
